requirements.txt
numpy==1.25.2
pandas==1.5.3
matplotlib==3.7.1
tqdm==4.65.0
flask==3.0.0
Flask-SQLAlchemy==3.1.1
flask-login==0.6.3